+++ src/database/models/film.js
@@ -24,7 +24,7 @@ module.exports = (sequelize, DataTypes) => {
   Film.init(
     {
       quantity: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       color: {
@@ -40,11 +40,11 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: true,
       },
       discount: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       extra: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       description: {
@@ -52,11 +52,11 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: true,
       },
       minimumRate: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       rate: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       status: {
