+++ src/database/models/weddingCard.js
@@ -33,7 +33,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: true,
       },
       quantity: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       color: {
@@ -61,15 +61,15 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: true,
       },
       rate: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true, 
       },
       discount: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       extra: {
-        type: DataTypes.BIGINT,
+        type: DataTypes.INTEGER,
         allowNull: true,
       },
       status: {
